+++ RELEASE_NOTES.md
@@ -1,3 +1,5 @@
 ## Changes
 
-- Change the exon diff to junction diff to make it more accurate.
+- Add ISOM_SAMPLE_CNT and ISOM_SAMPLE_FREQ attributes to output GTF for sample level statistics.
+
+